src/requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==4.2.0
blinker==1.9.0
cached-property==2.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==43.0.0
defusedxml==0.7.1
dnspython==2.8.0
exchangelib==5.0.1
fastapi==0.119.0
Flask==3.0.3
Flask-RESTful==0.3.10
h11==0.16.0
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpickle==4.1.1
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.3
mysql-connector-python==9.0.0
oauthlib==3.3.1
outcome==1.3.0.post0
psycopg2-binary==2.9.11
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyMySQL==1.1.1
PySocks==1.7.1
pyspnego==0.12.0
pytz==2025.2
repoze.lru==0.7
requests==2.32.3
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
Routes==2.5.1
selenium==4.36.0
sentry-sdk==2.13.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.48.0
stripe==13.0.1
timedelta==2020.12.3
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0